pyproject.toml
[tool.poetry]
name = "university-board-consensus"
version = "0.1.0"
description = "A multi-agent consensus system for university board decisions"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,!=3.9.7,<4.0"
python-dotenv = "^1.0.0"
networkx = "^3.1"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
streamlit = "^1.22.0"
plotly = "^5.14.1"
pandas = "^2.0.1"
tenacity = "^8.2.2"
numpy = "^1.24.3"
scikit-learn = "^1.6.0"
anthropic = "^0.42.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 